Move old loadTiles to tools
